feat(5-1): change structure of reviews folder
