+++ src/store/reviews/actions.js
@@ -1,0 +1,10 @@
+export const ActionType = {
+  LOAD_REVIEWS: `LOAD_REVIEWS`,
+};
+
+export const ActionCreator = {
+  loadReviews: (reviews) => ({
+    type: ActionType.LOAD_REVIEWS,
+    payload: reviews,
+  }),
+};
